feat(misc): 3824  - Move anonymization task  (#4646)
