+++ api/package.json
@@ -43,7 +43,6 @@
     "aws-sdk": "^2.1248.0",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.20.3",
-    "bson": "^4.7.2",
     "bullmq": "^5.10.0",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
