+++ api/src/anonymization/PlanDeTransport/ligneBus.js
@@ -1,5 +1,5 @@
-const { generateRandomName, generateBirthdate, generateRandomEmail, generateNewPhoneNumber } = require("../../utils/anonymise");
-const { anonymizeNonDeclaredFields } = require("../../utils/anonymise-model-fields");
+const { generateRandomName, generateBirthdate, generateRandomEmail, generateNewPhoneNumber } = require("../utils/anonymise");
+const { anonymizeNonDeclaredFields } = require("../utils/anonymise-model-fields");
 
 function anonymize(itemToAnonymize) {
   const whitelist = [
