+++ api/src/anonymization/cohesionCenter.js
@@ -1,4 +1,4 @@
-const { anonymizeNonDeclaredFields } = require("../utils/anonymise-model-fields");
+const { anonymizeNonDeclaredFields } = require("./utils/anonymise-model-fields");
 
 function anonymize(itemToAnonymize) {
   const whitelist = [
