+++ api/src/anonymization/young.js
@@ -1,6 +1,6 @@
-const { generateAddress, generateRandomName, generateRandomEmail, generateBirthdate, getYoungLocation, generateNewPhoneNumber, starify } = require("../utils/anonymise");
-const { anonymizeNonDeclaredFields } = require("../utils/anonymise-model-fields");
-const { v4: uuid } = require("uuid");
+const { generateAddress, generateRandomName, generateRandomEmail, generateBirthdate, getYoungLocation, generateNewPhoneNumber, starify } = require("./utils/anonymise");
+const { anonymizeNonDeclaredFields } = require("./utils/anonymise-model-fields");
+const { randomUUID } = require("crypto");
 
 function anonymize(itemToAnonymize) {
   const whitelist = [
@@ -399,8 +399,8 @@ function anonymize(itemToAnonymize) {
   item.forgotPasswordResetToken = "";
   item.invitationToken = "";
   item.phase3Token = "";
-  item.parent1Inscription2023Token = uuid();
-  item.parent2Inscription2023Token = uuid();
+  item.parent1Inscription2023Token = randomUUID();
+  item.parent2Inscription2023Token = randomUUID();
 
   return item;
 }
