--- api/src/scripts/anonymization/docker/start_anonymize_db.sh
@@ -1,11 +1,0 @@
-#!/bin/bash
-
-cd "$(dirname $0)/.."
-
-set -o pipefail
-set -e
-
-src_db_uri=$(../get_secrets.sh snu-production | jq -r ".MONGO_URL")
-dest_db_uri=$(../get_secrets.sh snu-staging | jq -r ".MONGO_URL")
-
-./anonymize_db.sh $src_db_uri $dest_db_uri
