+++ devops/anonymization/anonymize_db.sh
