+++ devops/anonymization/cc.env-example
@@ -1,0 +1,4 @@
+CC_DOCKERFILE=devops/anonymization/Dockerfile
+CC_TASK=true
+SOURCE_DATABASE_URI=<url>
+TARGET_DATABASE_URI=<url>
